extract to LikeIconButoon

+++ app/src/main/java/com/jorgediazp/kpopcalendar/home/common/data/liked/LikedSongsRepository.kt
@@ -19,7 +19,7 @@ class LikedSongsRepository @Inject constructor(
         return dataSource.getAllLikedSongs()
     }
 
-    override suspend fun deleteLikedSong(songId: Long): DataResult<Nothing> {
+    override suspend fun deleteLikedSong(songId: Int): DataResult<Nothing> {
         return dataSource.deleteLikedSong(songId)
     }
 }
