+++ app/src/main/java/com/jorgediazp/kpopcalendar/home/common/data/songs/SongsDataModelExtensions.kt
@@ -24,7 +24,7 @@ class SongsDataModelExtensions {
 
         fun HashMap<String, Any>.toDomainModel(): SongDomainModel {
             return SongDomainModel(
-                id = this[ID_FIELD] as Long,
+                id = (this[ID_FIELD] as Long).toInt(),
                 artist = this[ARTIST_FIELD] as String?,
                 artists = this[ARTISTS_FIELD] as List<String>?,
                 titleTrack = this[TITLE_TRACK_FIELD] as String?,
